requirements.txt
# Core Framework Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0
aiosqlite==0.19.0
redis==5.0.1
pydantic==2.5.2
pydantic-settings==2.1.0

# Cryptography and Security
pynacl==1.5.0
cryptography==41.0.8
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# GUI and Visualization
PyQt6==6.6.1
networkx==3.2.1
matplotlib==3.8.2
plotly==5.17.0
graphviz==0.20.1

# HTTP and Network
httpx==0.25.2
aiohttp==3.9.1
aiohttp[speedups]==3.9.1
dnspython==2.4.2
scapy==2.5.0

# Data Processing
pandas==2.1.4
numpy==1.25.2
Pillow==10.1.0
python-docx==1.1.0
PyPDF2==3.0.1

# Email and Communication
aiosmtplib==3.0.1
email-validator==2.1.0.post1

# Utilities and Logging
click==8.1.7
rich==13.7.0
loguru==0.7.2
python-multipart==0.0.6
jinja2==3.1.2
pyyaml==6.0.1
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development Tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0
